Add functions to read/write multiple pins at the same time
In some applications it is important that multiple writes or reads
happen at the same time and are not separated by multiple I2C
transactions.

Provide two new functions `write_multiple()` and `read_multiple()` which
allow performing multiple pin-state changes / multiple pin-state reads
in a single bus transaction (as far as the devices support this).