Allow glitchless output configuration
When configuring a pin as output, it was previously not possible to
glitchlessly bring the pin into the desired electrical state.  Change
the semantics of `into_output()` such that it always puts the pin into a
LOW state.  Add a new `into_output_high()` which always puts the pin
into a HIGH state.