Define our own mutex trait
Right now, we are still relying on `shared_bus::BusMutex`.  As
`shared-bus` has gotten obsolete with the release of `embedded-hal` 1.0,
it's time to drop our dependency on it.  To do this, we need a
replacement for `BusMutex`.  As there isn't really any other crate
providing something suitable, let's roll our own - exactly what
`shared-bus` did in its time as well.