Wrap errors for embedded-hal trait implementations
The current implementations of the `OutputPin`, `InputPin` and `StatefulOutputPin`
traits require that `PortDriver::Error` implements `embedded_hal::digital::Error`,
which is never going to be the case. Thus, these traits are never usable.

This change wraps the error type used in `ErrorKind::Error` in a wrapper struct
which implements `embedded_hal::digital::Error`. Like this all upstream errors that
are `core::fmt::Debug` should be compatible with `embedded-hal`.